Remove int32 cast from StructureExporter.
PiperOrigin-RevId: 253281917

+++ morph_net/tools/structure_exporter.py
@@ -70,8 +70,7 @@ class StructureExporter(object):
         tf.logging.warning('No regularizer found for: %s', op.name)
         continue
 
-      self._tensors[rename_fn(op.name)] = tf.cast(op_regularizer.alive_vector,
-                                                  tf.int32)
+      self._tensors[rename_fn(op.name)] = op_regularizer.alive_vector
 
   @property
   def tensors(self):
